moved the state for deck and setdeck to deckscreen lower down in hierarchy for ease of prop handling and passing down to other components

+++ src/deck-screen/Card.js
@@ -11,20 +11,28 @@ import { deleteCard } from "../utils/api";
 // has edit button (goes to Edit Card screen)
 // delete button (with warning message. )
 
-function Card({ card, deleteCardById }) {
+function Card({ card }) {
   const { url } = useRouteMatch();
-
-
+  console.log("Card Data:", card)
+  console.log({url});
 
   const handleDelete = async () => {
     const abortController = new AbortController();
-    await deleteCard(card.id, abortController.signal);
-    // console.log("Deleting card with ID:", card.id)
-    // deleteCardById(card.id);
-
+    const result = window.confirm(`Do you want to delete this card? \n \nYou will not be able to recover it.`);
+    if (result) { 
+        await deleteCard(card.id, abortController.signal);
+    }
   }
 
-  console.log("Card Id:", card.id);
+//   const handleDelete = async () => {
+//     const abortController = new AbortController();
+//     await deleteCard(card.id, abortController.signal);
+//     // console.log("Deleting card with ID:", card.id)
+//     // deleteCardById(card.id);
+
+//   }
+
+//   console.log("Card Id:", card.id);
 
   return (
     <div className="card mx-auto">
@@ -37,7 +45,7 @@ function Card({ card, deleteCardById }) {
             </div>
             <div className="d-flex justify-content-end">
                 
-              <Link to={`{${url}/cards/${card.id}/edit`} className="btn btn-secondary mr-2">
+              <Link to={`${url}/cards/${card.id}/edit`} className="btn btn-secondary mr-2">
                 <i className="bi bi-pen"></i> Edit
               </Link>
